Use linear any angle validation algorithm.
The old one is O(N^2). But to enable linear algorithm, an
List<Set<Integer>> has to been copied, which outweight the gain from
O(N^2) to O(N) in any angle validation.

One drawback of current iterative algorithm is it is kind of Breadth
First Visiting instead of Depth First Visiting in the recursive
implementation. The BFS requires the status to be deep copied while
DFS does not because each DFS continuous visiting status is maintained
by backtracking. And the deep copying of status is expensive.

So the iterative path is abandoned.